win: move windows-mmap to its own file
